requirements.txt
Django>=4.2.0
groq>=0.4.0
python-dotenv>=1.0.0

# Django REST framework for API
djangorestframework>=3.14.0
# JWT authentication for API endpoints
djangorestframework-simplejwt>=5.3.0
# CORS headers for frontend connection
django-cors-headers>=4.0.0
# Password validation and security
django-extensions>=3.2.0

# Enhanced RAG and Vector Search
sentence-transformers>=2.5.1
faiss-cpu>=1.7.4
numpy>=1.24.0

# Real-time speech recognition and audio processing
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
# Audio format conversion for WebM to WAV
pydub>=0.25.1

# User agent parsing for session tracking
user-agents>=2.2.0

# Optional speech service providers (install as needed)
# azure-cognitiveservices-speech>=1.31.0  # For Azure Speech Service
# google-cloud-speech>=2.21.0            # For Google Cloud Speech 